requirements.txt
flask

flask-sqlalchemy

psycopg2

python-dotenv
